third_party/blink/web_tests/http/tests/security/isolatedWorld/isolated-world-fetch-csp-expected.txt

CONSOLE MESSAGE: Testing main world. Request should be blocked by main world CSP.
CONSOLE ERROR: Connecting to 'http://127.0.0.1:8000/security/isolatedWorld/resources/access_control_allow_origin.php' violates the following Content Security Policy directive: "connect-src 'none'". The action has been blocked.
CONSOLE ERROR: Fetch API cannot load http://127.0.0.1:8000/security/isolatedWorld/resources/access_control_allow_origin.php. Refused to connect because it violates the document's Content Security Policy.
CONSOLE MESSAGE: PASS: Request blocked by CSP as expected.
CONSOLE MESSAGE: Testing isolated world with no csp. Request should be blocked by main world CSP.
CONSOLE ERROR: Connecting to 'http://127.0.0.1:8000/security/isolatedWorld/resources/access_control_allow_origin.php' violates the following Content Security Policy directive: "connect-src 'none'". The action has been blocked.
CONSOLE ERROR: Fetch API cannot load http://127.0.0.1:8000/security/isolatedWorld/resources/access_control_allow_origin.php. Refused to connect because it violates the document's Content Security Policy.
CONSOLE MESSAGE: PASS: Request blocked by CSP as expected.
CONSOLE MESSAGE: Testing isolated world with permissive csp.
CONSOLE MESSAGE: PASS: Request succeeded as expected.
CONSOLE MESSAGE: Testing fetch redirect in isolated world with permissive csp.
CONSOLE MESSAGE: PASS: Request succeeded as expected.
CONSOLE MESSAGE: Testing isolated world with strict csp.
CONSOLE ERROR: Connecting to 'http://127.0.0.1:8000/security/isolatedWorld/resources/access_control_allow_origin.php' violates the following Content Security Policy directive: "connect-src 'self'". The action has been blocked.
CONSOLE ERROR: Fetch API cannot load http://127.0.0.1:8000/security/isolatedWorld/resources/access_control_allow_origin.php. Refused to connect because it violates the document's Content Security Policy.
CONSOLE MESSAGE: PASS: Request blocked by CSP as expected.
CONSOLE MESSAGE: Testing fetch redirect in isolated world with strict csp.
CONSOLE ERROR: Connecting to 'http://127.0.0.1:8000/resources/redirect.php?url=http://127.0.0.1:8000/security/isolatedWorld/resources/access_control_allow_origin.php&cors_allow_origin=*&delay=100' violates the following Content Security Policy directive: "connect-src 'self'". The action has been blocked.
CONSOLE ERROR: Fetch API cannot load http://127.0.0.1:8000/resources/redirect.php?url=http://127.0.0.1:8000/security/isolatedWorld/resources/access_control_allow_origin.php&cors_allow_origin=*&delay=100. Refused to connect because it violates the document's Content Security Policy.
CONSOLE MESSAGE: PASS: Request blocked by CSP as expected.
This tests the interaction of the fetch API run in the isolated world with the isolated world CSP.
